Refactoring implemented: naming was updated, file structure updated
