+++ YetAnotherWeather/CoreLayer/Services/CurrentWeatherService/Models/ConditionModel.swift
@@ -1,0 +1,13 @@
+//
+//  ConditionModel.swift
+//  YetAnotherWeather
+//
+//  Created by Ivan Grebenyuk on 24.09.2024.
+//
+
+import Foundation
+
+struct Condition: Codable {
+    let text: String
+    let icon: String
+}
