--- YetAnotherWeather/UserStories/CurrentWeatherList/Core/Models/ConditionModel.swift
@@ -1,13 +1,0 @@
-//
-//  ConditionModel.swift
-//  YetAnotherWeather
-//
-//  Created by Иван Гребенюк on 22.09.2024.
-//
-
-import Foundation
-
-struct Condition: Codable {
-    let text: String
-    let icon: String
-}
