+++ YetAnotherWeather/UserStories/Presentation/CurrentWeatherList/CurrentWeatherListAssembly.swift
@@ -1,0 +1,44 @@
+//
+//  CurrentWeatherListAssembly.swift
+//  YetAnotherWeather
+//
+//  Created by Ivan Grebenyuk on 13.05.2024.
+//
+
+import UIKit
+
+class CurrentWeatherListAssembly {
+    
+    // Dependencies
+    let currentWeatherService: ICurrentWeatherService
+    let searchResultAssembly: SearchResultsAssembly
+    
+    init(
+        weatherNetworkService: ICurrentWeatherService,
+        searchResultAssembly: SearchResultsAssembly
+    ) {
+        self.currentWeatherService = weatherNetworkService
+        self.searchResultAssembly = searchResultAssembly
+    }
+    
+    func assemble(output: CurrentWeatherListOutput?) -> UIViewController {
+        
+        let presenter = CurrentWeatherListPresenter(
+            currentWeatherService: currentWeatherService,
+            output: output
+        )
+        
+        let searchResultsViewController = searchResultAssembly.assemble(
+            output: presenter
+        )
+        
+        let viewController = CurrentWeatherListViewController(
+            resultsViewController: searchResultsViewController,
+            presenter: presenter
+        )
+        
+        presenter.view = viewController
+        return viewController
+    }
+}
+
