+++ YetAnotherWeather/UserStories/Presentation/CurrentWeatherList/Submodules/SearchResults/Views/SearchResultCellView.swift
