+++ YetAnotherWeather/UserStories/Presentation/CurrentWeatherList/Views/CurrentWeatherCellView.swift
