--- YetAnotherWeather/UserStories/WeatherDetails/Core/Models/Forecast.swift
@@ -1,22 +1,0 @@
-//
-//  Forecast.swift
-//  YetAnotherWeather
-//
-//  Created by Ivan Grebenyuk on 30.07.2024.
-//
-
-import Foundation
-
-struct Forecast {
-    
-    struct ForecastDay {
-        let date: Int
-        let maxTemp: Double
-        let minTemp: Double
-        let avgTemp: Double
-        let condition: Condition
-    }
-    
-    let currentWeather: CurrentWeatherModel
-    let forecastDays: [ForecastDay]
-}
